requirements.txt
psutil
openpyxl
matplotlib
pyserial